requirements.txt
fastapi
uvicorn
opencv-python
numpy
pyyaml
ultralytics
